feat: enhance transaction handling in protocol buffers
- Add `BlockNumber` field to `Transaction` message in protocol buffer definition
- Update `BlockNumber` type in `TransactionDAO` from `int` to `int64`
- Change `Value` type in `TransactionDAO` from `float64` to `int64`
- Implement `Create` method in `TransactionRepoImpl` to insert transactions into the database
- Remove TODO comment from the `Create` method implementation

+++ internal/domain/core/infra/storage/pg/transaction_repo_dao.go
@@ -7,11 +7,11 @@ import (
 // TransactionDAO represents the transactions table
 type TransactionDAO struct {
 	TxHash      string    `db:"tx_hash"`      // TransactionDAO hash
-	BlockNumber int       `db:"block_number"` // Block number
+	BlockNumber int64     `db:"block_number"` // Block number
 	Timestamp   time.Time `db:"timestamp"`    // TransactionDAO timestamp
 	FromAddress string    `db:"from_address"` // Sender address
 	ToAddress   string    `db:"to_address"`   // Receiver address
-	Value       float64   `db:"value"`        // TransactionDAO value
+	Value       int64     `db:"value"`        // TransactionDAO value
 	GasUsed     int       `db:"gas_used"`     // Gas used
 	GasPrice    float64   `db:"gas_price"`    // Gas price
 	Status      bool      `db:"status"`       // TransactionDAO status (success or failure)
